Reorganized task stuff so task functions are in webapp code
